- load the JACScript and create the JAntiCaptcha object just, if needed - renamed UTILITIES to Utilities
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@6054 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/captcha/specials/MegaUpload.java
@@ -54,7 +54,7 @@ public class MegaUpload {
 
         ArrayList<Letter> ret = new ArrayList<Letter>();
         for (int i = 0; i < 4; i++) {
-            int averageWidth = Math.min(captcha.getWidth(), (int) (captcha.getWidth() / (4 - i)) + 20);
+            int averageWidth = Math.min(captcha.getWidth(), (captcha.getWidth() / (4 - i)) + 20);
             Letter first = new Letter(averageWidth, captcha.getHeight());
 
             first.setId(i);
